code improvements

+++ src/main/java/me/offeex/bloomware/client/module/modules/hud/ItemDurability.java
@@ -17,10 +17,10 @@ public class ItemDurability extends Module {
         if (isEnabled()) {
             assert mc.player != null;
             if (!mc.player.getMainHandStack().isEmpty() || !mc.player.getMainHandStack().isDamageable()) {
-                width = (int) (Bloomware.sFontRenderer.getStringWidth("Durability: -", Bloomware.sFontRenderer.getFontsize()));
+                width = Bloomware.sFontRenderer.getStringWidth("Durability: -", Bloomware.sFontRenderer.getFontsize());
                 text = "Durability: -";
             } else {
-                width = (int) (Bloomware.sFontRenderer.getStringWidth("Durability: " + (mc.player.getMainHandStack().getMaxDamage() - mc.player.getMainHandStack().getDamage()), Bloomware.sFontRenderer.getFontsize()));
+                width = Bloomware.sFontRenderer.getStringWidth("Durability: " + (mc.player.getMainHandStack().getMaxDamage() - mc.player.getMainHandStack().getDamage()), Bloomware.sFontRenderer.getFontsize());
                 text = "Durability: " + (mc.player.getMainHandStack().getMaxDamage() - mc.player.getMainHandStack().getDamage());
             }
 //            if (ClickGui.getCurrentScreen() == 2) {
